+++ physicsnemo/diffusion/noise_schedulers/__init__.py
@@ -14,16 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .noise_schedulers import (  # noqa: F401
-    EDMLogUniformNoiseScheduler,
-    EDMNoiseScheduler,
-    IDDPMNoiseScheduler,
-    LinearGaussianNoiseScheduler,
-    NoiseScheduler,
-    StudentTEDMNoiseScheduler,
-    VENoiseScheduler,
-    VPNoiseScheduler,
-)
+from .base import NoiseScheduler  # noqa: F401
+from .edm import EDMNoiseScheduler  # noqa: F401
+from .edm_log_uniform import EDMLogUniformNoiseScheduler  # noqa: F401
+from .iddpm import IDDPMNoiseScheduler  # noqa: F401
+from .linear_gaussian import LinearGaussianNoiseScheduler  # noqa: F401
+from .student_t_edm import StudentTEDMNoiseScheduler  # noqa: F401
+from .ve import VENoiseScheduler  # noqa: F401
+from .vp import VPNoiseScheduler  # noqa: F401
 
 
 def __getattr__(name: str):
