+++ physicsnemo/diffusion/samplers/__init__.py
@@ -14,13 +14,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .base import Solver  # noqa: F401
+from .edm_stochastic_euler import EDMStochasticEulerSolver  # noqa: F401
+from .edm_stochastic_heun import EDMStochasticHeunSolver  # noqa: F401
+from .euler import EulerSolver  # noqa: F401
+from .heun import HeunSolver  # noqa: F401
 from .legacy_deterministic_sampler import deterministic_sampler  # noqa: F401
 from .legacy_stochastic_sampler import stochastic_sampler  # noqa: F401
 from .samplers import sample  # noqa: F401
-from .solvers import (  # noqa: F401
-    EDMStochasticEulerSolver,
-    EDMStochasticHeunSolver,
-    EulerSolver,
-    HeunSolver,
-    Solver,
-)
